email-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.24.1
python-multipart==0.0.6
pydantic==2.3.0
jinja2==3.1.2
aiosmtplib==2.0.2
google-generativeai==0.3.1
python-dotenv==1.0.0
email-validator==2.0.0.post2